Optimize winner method and its dependent methods
winner was not able to handle a lot of samples at a time, therefore the
methods calling it were processing one element at a time. This is very
inefficient since it implied loading/unloading GPU memory everytime, as
well as limiting the possible parallelization to just one sample.
This patch, makes it possible to handle an arbitrary number of samples,
processing them in chunks, like train does. Methods calling winner will
now call it with the full data samples and then iterate over the result.